fix: after moving components
